all /api/items/ pages
