+++ website/api/items/unvoteItem.ts
@@ -11,7 +11,7 @@ export default function unvoteItem(itemId, callback) {
             },
             {
                 withCredentials: true,
-            },
+            }
         )
         .then(function (response) {
             callback(response.data);
