+++ website/api/items/upvoteItem.ts
@@ -11,7 +11,7 @@ export default function upvoteItem(itemId, callback) {
             },
             {
                 withCredentials: true,
-            },
+            }
         )
         .then(function (response) {
             callback(response.data);
