+++ website/pages/edit-item.js
@@ -6,8 +6,8 @@ import Header from "../components/header.js";
 import Footer from "../components/footer.js";
 import HeadMetadata from "../components/HeadMetadata";
 
-import getEditItemPageData from "../api/items/getEditItemPageData.js";
-import editItem from "../api/items/editItem.js";
+import getEditItemPageData from "../api/items/getEditItemPageData";
+import editItem from "../api/items/editItem";
 
 import renderCreatedTime from "../utils/renderCreatedTime";
