+++ website/pages/favorites.js
@@ -6,7 +6,7 @@ import HeadMetadata from "../components/HeadMetadata";
 import ItemsList from "../components/itemsList.js";
 import CommentsList from "../components/commentsList.js";
 
-import getUserFavoritedItemsByPage from "../api/items/getUserFavoritedItemsByPage.js";
+import getUserFavoritedItemsByPage from "../api/items/getUserFavoritedItemsByPage";
 import getUserFavoritedCommentsByPage from "../api/comments/getUserFavoritedCommentsByPage.js";
 
 export default function Favorites({
