+++ website/pages/hidden.js
@@ -3,7 +3,7 @@ import Footer from "../components/footer.js";
 import HeadMetadata from "../components/HeadMetadata";
 import ItemsList from "../components/itemsList.js";
 
-import getUserHiddenItemsByPage from "../api/items/getUserHiddenItemsByPage.js";
+import getUserHiddenItemsByPage from "../api/items/getUserHiddenItemsByPage";
 
 export default function Hidden({ items, authUserData, page, isMore, getDataError, goToString }) {
     return (
