+++ website/pages/past.js
@@ -6,7 +6,7 @@ import Footer from "../components/footer.js";
 import HeadMetadata from "../components/HeadMetadata";
 import ItemsList from "../components/itemsList.js";
 
-import getRankedItemsByDay from "../api/items/getRankedItemsByDay.js";
+import getRankedItemsByDay from "../api/items/getRankedItemsByDay";
 
 export default function Past({ items, authUserData, day, page, isMore, getDataError, invalidDateError, goToString }) {
     const renderGoBackwardLinks = () => {
