+++ website/pages/upvoted.js
@@ -4,7 +4,7 @@ import HeadMetadata from "../components/HeadMetadata";
 import ItemsList from "../components/itemsList.js";
 import CommentsList from "../components/commentsList.js";
 
-import getUserUpvotedItemsByPage from "../api/items/getUserUpvotedItemsByPage.js";
+import getUserUpvotedItemsByPage from "../api/items/getUserUpvotedItemsByPage";
 import getUserUpvotedCommentsByPage from "../api/comments/getUserUpvotedCommentsByPage.js";
 
 export default function Upvoted({
